+++ packages/vendure-plugin-frequently-bought-together/src/custom-fields-types.ts
@@ -7,5 +7,10 @@ import {
 declare module '@vendure/core' {
   interface CustomProductFields {
     frequentlyBoughtWith?: Product[];
+    /**
+     * Stringified JSON text of Support[]
+     * This field holds the support level per product, so that we can return a sorted list of products based on their support level
+     */
+    frequentlyBoughtWithSupport?: string;
   }
 }