+++ packages/vendure-plugin-frequently-bought-together/CHANGELOG.md
@@ -1,3 +1,9 @@
+# 1.0.0 (2024-12-12)
+
+- Implemented FPGrowth algorithm for defining frequently bought together relations
+- Added experiment mode to test the support level needed
+- Return sorted list of related products based on support level
+
 # 0.0.2 (2024-12-03)
 
 - Patched compatibility range to >= 2.2.0